requirements.txt
# --- Runtime ---
Flask==3.0.3
gunicorn==21.2.0
pymongo[srv]==4.6.3
APScheduler==3.10.4
pytz==2024.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
blinker==1.8.2
# Optional validation helpers
phonenumberslite==8.13.47

# --- Templating/Linting for Jinja files (dev-time; safe at runtime) ---
djlint==1.34.1

# --- Dev/Test ---
pytest==8.3.2
pytest-cov==5.0.0
coverage==7.6.4
black==24.8.0
isort==5.13.2
ruff==0.6.5
mypy==1.11.2
bandit==1.7.9
pydocstyle==6.3.0
freezegun==1.5.1
Faker==26.0.0
